+++ cloud/cloud.go
@@ -146,6 +146,8 @@ const (
 	//elasticbeanstalk
 	Application string = "application"
 	Environment string = "environment"
+	//msk
+	KafkaCluster string = "kafkacluster"
 	//cognito. Two APIs: user pools on cognito-idp, identity pools on cognito-identity.
 	UserPool     string = "userpool"
 	IdentityPool string = "identitypool"